File for testing GUI and game functions.
This file contains lots of different GUI settings.
- Frames
- Buttons
- Drop down menu
- Text input
- Labels
- Canvas

Within the main canvas the program simulates collisions between the
large sun objects and the smaller planet object. Both detect the
boundaries of the screen and bounce off. There is also a player object
which moves with momentum using the arrow keys. If it hits the sun it
will be broken.