Input for time changed, based on user feedback
Users where just inputting for seconds and entering nothing for minutes
when asked to input 30 seconds. Thus changed it so it now works. Finally
user can now enter 0.5 into minutes or seconds and that will still work
for the time input. For example 0.5 in minutes input will make program
run for 30 seconds.